# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BlindPass SPS API
  version: 0.1.0
  description: |
    Maintained OpenAPI snapshot for the stable BlindPass hosted and local-development API surfaces.
    This document is derived from the route schemas in `packages/sps-server/src/routes`.
servers:
  - url: http://127.0.0.1:3100
    description: Local development
  - url: https://sps.atas.tech
    description: Hosted production
...
paths:
...
  /api/v2/secret/request:
    post:
      tags: [Secrets]
      summary: Create a human-to-agent secret request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [public_key, description]
              properties:
                public_key:
                  type: string
                description:
                  type: string
                  minLength: 1
                  maxLength: 512
      responses:
        '201':
          description: Secret request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                  confirmation_code:
                    type: string
                  secret_url:
                    type: string
                    format: uri
                required: [request_id, confirmation_code, secret_url]
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
  /api/v2/secret/metadata/{id}:
    get:
      tags: [Secrets]
      summary: Read browser-side metadata for a signed secret request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: sig
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public key and confirmation metadata for the browser UI
          content:
            application/json:
              schema:
                type: object
                properties:
                  public_key:
                    type: string
                  description:
                    type: string
                  confirmation_code:
                    type: string
                  expiry:
                    type: integer
                required: [public_key, description, confirmation_code, expiry]
        '403':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
  /api/v2/secret/submit/{id}:
    post:
      tags: [Secrets]
      summary: Submit an encrypted secret payload from the browser UI
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: sig
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enc, ciphertext]
              properties:
                enc:
                  type: string
                ciphertext:
                  type: string
      responses:
        '201':
          description: Ciphertext accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: submitted
                required: [status]
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
  /api/v2/secret/status/{id}:
    get:
      tags: [Secrets]
      summary: Poll the status of a secret request
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current request state
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, submitted, expired]
                required: [status]
        '401':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
  /api/v2/secret/retrieve/{id}:
    get:
      tags: [Secrets]
      summary: Retrieve and delete a submitted secret payload
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Encrypted payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  enc:
                    type: string
                  ciphertext:
                    type: string
                required: [enc, ciphertext]
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentApiKey:
      type: apiKey
      in: header
      name: x-agent-api-key
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        issues:
          type: array
          items:
            type: object
      required: [error]
...